refactor: Renamed `landing-page` folder to `website`
